Properly create a single user and traverse the info back to main()
HUGE THANKS TO Lewisk3!  Without your help, I think I would have been doomed for a long awhile as the pointers continued to erode all sanity!

It is now possible to properly create a single user within the Linked-List and update the list properly back to main().  This will allow further expandability as we continue to grow the code.