Corrected logic error
I don't know what I was thinking, but the previous implementation - would never work or is not practical.  When a RandomType is incorrect, we should throw _a_ value in the first index, but at the next - we should terminate it properly.  More I think about this, the more I think this could be grossly incorrect.  I might redo this function later or thrash it if it is not necessary.