[UNTESTED] Added functionality to scan user DB for a specific account
NOTE: THIS IS VERY WIP AND UNTESTED!

Added the ability to scan through the User's Linked-List to find a specific account.  If we find the user account and the authentication challenge was successful, than the UserCard is returned containing the user's information.  Otherwise, a failure has occurred as the user may not have typed in the correct key or no account exists within the Linked-List.

This is our pathway to the Login System.