Progress Update: Welcome Procedure
This is a very incomplete thought right now, but should be uploaded as a Progress Update so the progress is not lost.
The idea is to prepare a small menu for when the user connects to the store.  The user will have their choice to: leave, register, or login to a hard-coded account.