Dropped the ability to filter and search features
Yes, these features are indeed possible to implement, but it will take more time to properly implement than what is offered.  This project has a massive time constraint - a few more weeks would secure the ability to add in those features, but there currently not enough time as it is.